[FIX] fix mask

+++ network/transformer_graph/main.py
@@ -114,11 +114,11 @@ class Trainer:
         loss = F.binary_cross_entropy(torch.sigmoid(pred[:, :-1]), trg[:, 1:], reduction='none')
 
         # pad_idx에 해당하는 레이블을 무시하기 위한 mask 생성
-        pad_mask = get_pad_mask(trg[:, 1:, 0], pad_idx=self.pad_idx).float()
+        pad_mask = get_pad_mask(trg[:, 1:, 0], pad_idx=self.pad_idx)
         sub_mask = get_subsequent_mask(trg[:, :, 0])
         mask = pad_mask.unsqueeze(-1).expand(-1, -1, loss.shape[2]) & sub_mask[:, :sub_mask.shape[1] - 1, :]
         # mask 적용
-        masked_loss = loss * mask
+        masked_loss = loss * mask.float()
         # 손실의 평균 반환
         return masked_loss.mean()
